Fix v2cpp runtime

+++ Live2D/V2/Framework/L2DEyeBlink.cpp
@@ -12,12 +12,12 @@ void L2DEyeBlink::setEyeMotion(int closeMs, int closedMs, int openMs) {
     mOpeningMs = static_cast<float>(openMs);
 }
 void L2DEyeBlink::updateParam(ALive2DModel* model) {
-    float now = static_cast<float>(UtSystem::getUserTimeMSec());
+    double now = UtSystem::getUserTimeMSec();
     mCurrentTime = now;
 
     switch (mState) {
     case FIRST:
-        mNextBlinkTime = now + static_cast<float>(rand() % (2 * (int)mBlinkIntervalMs));
+        mNextBlinkTime = now + static_cast<double>(rand() % (2 * (int)mBlinkIntervalMs));
         mState = INTERVAL;
         break;
     case INTERVAL:
@@ -27,8 +27,8 @@ void L2DEyeBlink::updateParam(ALive2DModel* model) {
         }
         break;
     case CLOSING: {
-        float elapsed = now - mStateStartTime;
-        float v = 1.0f - (elapsed / mClosingMs);
+        double elapsed = now - mStateStartTime;
+        float v = static_cast<float>(1.0 - (elapsed / mClosingMs));
         if (v <= 0) { v = 0; mState = CLOSED; mStateStartTime = now; }
         model->setParamFloat(model->getModelContext()->getParamIndex(&Id::getID(mEyeLId)), v);
         model->setParamFloat(model->getModelContext()->getParamIndex(&Id::getID(mEyeRId)), v);
@@ -37,15 +37,15 @@ void L2DEyeBlink::updateParam(ALive2DModel* model) {
     case CLOSED: {
         model->setParamFloat(model->getModelContext()->getParamIndex(&Id::getID(mEyeLId)), 0);
         model->setParamFloat(model->getModelContext()->getParamIndex(&Id::getID(mEyeRId)), 0);
-        float elapsed = now - mStateStartTime;
+        double elapsed = now - mStateStartTime;
         if (elapsed >= mClosedMs) { mState = OPENING; mStateStartTime = now; }
         break;
     }
     case OPENING: {
-        float elapsed = now - mStateStartTime;
-        float v = elapsed / mOpeningMs;
+        double elapsed = now - mStateStartTime;
+        float v = static_cast<float>(elapsed / mOpeningMs);
         if (v >= 1.0f) { v = 1.0f; mState = INTERVAL;
-            mNextBlinkTime = now + mBlinkIntervalMs + static_cast<float>(rand() % (int)mBlinkIntervalMs); }
+            mNextBlinkTime = now + mBlinkIntervalMs + static_cast<double>(rand() % (int)mBlinkIntervalMs); }
         model->setParamFloat(model->getModelContext()->getParamIndex(&Id::getID(mEyeLId)), v);
         model->setParamFloat(model->getModelContext()->getParamIndex(&Id::getID(mEyeRId)), v);
         break;
